Encode networkid as a u64.
Conflicts:
	ethcore/src/types/transaction.rs

+++ ethcore/src/client/client.rs
@@ -1184,7 +1184,7 @@ impl BlockChainClient for Client {
 		self.miner.pending_transactions(self.chain.read().best_block_number())
 	}
 
-	fn signing_network_id(&self) -> Option<u8> {
+	fn signing_network_id(&self) -> Option<u64> {
 		self.engine.signing_network_id(&self.latest_env_info())
 	}
